Rename to linearization_merge for consistency with other modules
