+++ src/spyglass/linearization/__init__.py
@@ -1,1 +1,1 @@
-from spyglass.linearization.merge import LinearizedPositionOutput
+from spyglass.linearization.linearization_merge import LinearizedPositionOutput
